Fix and test Series.next_occurrences()
For monthly meetups, the next_occurrences(since=...) method excluded
the `since` argument's month. This was there to prevent the situation
that a meetup would be held early (e.g. on Monday), but the regular
date would still show up in the predictions.

Fix and the logic to plan from the last scheduled meetup, or from
the next month if the meetu is monthly.
Add tests.